add witness serialization, fix tests, add zkutil to test
