+++ zokrates_circom/Cargo.toml
@@ -12,3 +12,4 @@ byteorder = "1.4.3"
 
 [dev-dependencies]
 pretty_assertions = "1.2.1"
+zkutil = "0.5.0"
